requirements.txt
colorama
pywhatkit
opencv-python
numpy
pillow
setproctitle